Added tests for DateKMM
