+++ shared/swiftpackage/shared.xcframework/Info.plist
@@ -5,8 +5,6 @@
 	<key>AvailableLibraries</key>
 	<array>
 		<dict>
-			<key>DebugSymbolsPath</key>
-			<string>dSYMs</string>
 			<key>LibraryIdentifier</key>
 			<string>ios-x86_64-simulator</string>
 			<key>LibraryPath</key>
@@ -21,8 +19,6 @@
 			<string>simulator</string>
 		</dict>
 		<dict>
-			<key>DebugSymbolsPath</key>
-			<string>dSYMs</string>
 			<key>LibraryIdentifier</key>
 			<string>ios-arm64</string>
 			<key>LibraryPath</key>
